Get rid of the Singleton Pattern.
Allows for multiple instances of Cognition to be running in one ruby
namespace, each managing their own plugins. Useful if you are running
several chat endpoints under one sinatra or rails app.
