+++ test/test_plugin.rb
@@ -4,7 +4,8 @@ require 'cognition'
 class PluginTest < Minitest::Test
   def test_sets_matchers
     require_relative "fixtures/hello"
-    hello = Hello.new
+    bot = Cognition::Bot.new
+    hello = Hello.new(bot)
 
     assert_equal 1, hello.matchers.count
   end
